ML/requirements.txt
scikit-learn
joblib
numpy
pandas
python-dotenv
matplotlib
seaborn
jupyterlab
ipykernel
